refactor: extract boosters opening and buying from lobby room logic
