+++ app/public/src/game/lobby-logic.ts
@@ -10,7 +10,6 @@ import {
 import LobbyState from "../../../rooms/states/lobby-state"
 import PreparationState from "../../../rooms/states/preparation-state"
 import { Transfer } from "../../../types"
-import type { Booster } from "../../../types/Booster"
 import { CloseCodes, CloseCodesMessages } from "../../../types/enum/CloseCodes"
 import { ConnectionStatus } from "../../../types/enum/ConnectionStatus"
 import type { NonFunctionPropNames } from "../../../types/HelperTypes"
@@ -42,11 +41,11 @@ import {
   removeTournament,
   removeTournamentBracket,
   resetLobby,
-  setBoosterContent,
   setCcu,
   setSearchedUser,
   updateTournament
 } from "../stores/LobbyStore"
+import { resetBoosters } from "../stores/BoostersStore"
 import {
   setConnectionStatus,
   setErrorAlertMessage,
@@ -119,6 +118,7 @@ export async function joinLobbyRoom(
                 dispatch(setErrorAlertMessage(t(`errors.${errorMessage}`)))
               }
               dispatch(resetLobby())
+              dispatch(resetBoosters())
               navigate("/")
             }
           })
@@ -270,13 +270,6 @@ export async function joinLobbyRoom(
             dispatch(setSearchedUser(user))
           )
 
-          room.onMessage(
-            Transfer.BOOSTER_CONTENT,
-            (boosterContent: Booster) => {
-              dispatch(setBoosterContent(boosterContent))
-            }
-          )
-
           joinLobby(room) // lobby room is now fully initialized and accessible
           resolve(room)
         } catch (error) {
@@ -320,6 +313,7 @@ export async function joinExistingPreparationRoom(
       joinPreparation(room, 30)
       leaveRoom("lobby")
       dispatch(resetLobby())
+      dispatch(resetBoosters())
       navigate("/preparation")
     }
   } catch (error: any) {
