+++ app/public/src/pages/component/main-sidebar/main-sidebar.tsx
@@ -192,7 +192,7 @@ export function MainSidebar(props: MainSidebarProps) {
             {t("collection")}
           </NavLink>
         )}
-        {page === "main_lobby" && profileLevel >= GADGETS.bag.levelRequired && (
+        {(page === "main_lobby" || page === "preparation") && profileLevel >= GADGETS.bag.levelRequired && (
           <NavLink
             location="booster"
             svg="booster"
@@ -246,14 +246,14 @@ export function MainSidebar(props: MainSidebarProps) {
         {((!GADGETS.synergy_wheel.disabled &&
           profileLevel >= GADGETS.synergy_wheel.levelRequired) ||
           profile?.role === Role.ADMIN) && (
-          <NavLink
-            svg="synergy-wheel"
-            location="synergy-wheel"
-            handleClick={changeModal}
-          >
-            {t("gadget.synergy_wheel")}
-          </NavLink>
-        )}
+            <NavLink
+              svg="synergy-wheel"
+              location="synergy-wheel"
+              handleClick={changeModal}
+            >
+              {t("gadget.synergy_wheel")}
+            </NavLink>
+          )}
 
         {page !== "game" &&
           ((!GADGETS.bot_builder.disabled &&
@@ -276,14 +276,14 @@ export function MainSidebar(props: MainSidebarProps) {
         {((!GADGETS.tier_list_maker.disabled &&
           profileLevel >= GADGETS.tier_list_maker.levelRequired) ||
           profile?.role === Role.ADMIN) && (
-          <NavLink
-            svg="tier-list"
-            location="tier-list"
-            handleClick={changeModal}
-          >
-            {t("gadget.tier_list_maker")}
-          </NavLink>
-        )}
+            <NavLink
+              svg="tier-list"
+              location="tier-list"
+              handleClick={changeModal}
+            >
+              {t("gadget.tier_list_maker")}
+            </NavLink>
+          )}
 
         {page !== "game" &&
           (profile?.role === Role.MODERATOR ||
