+++ app/public/src/stores/LobbyStore.ts
@@ -6,7 +6,6 @@ import {
   TournamentSchema
 } from "../../../models/colyseus-models/tournament"
 import { IChatV2, IGameMetadata, IPreparationMetadata } from "../../../types"
-import type { Booster } from "../../../types/Booster"
 import { Language } from "../../../types/enum/Language"
 import {
   ILeaderboardBotInfo,
@@ -26,8 +25,6 @@ export interface IUserLobbyState {
   tabIndex: number
   preparationRooms: RoomAvailable[]
   gameRooms: RoomAvailable[]
-  boosterContent: Booster
-  lastBoostersOpened: Booster[]
   language: Language
   tournaments: TournamentSchema[]
   ccu: number
@@ -35,8 +32,6 @@ export interface IUserLobbyState {
 
 const initialState: IUserLobbyState = {
   language: Language.en,
-  boosterContent: [],
-  lastBoostersOpened: [],
   messages: [],
   leaderboard: [],
   botLeaderboard: [],
@@ -118,13 +113,6 @@ export const lobbySlice = createSlice({
       action: PayloadAction<IUserMetadataClient | undefined>
     ) => {
       state.searchedUser = action.payload
-    },
-    setBoosterContent: (state, action: PayloadAction<Booster>) => {
-      state.boosterContent = action.payload
-      state.lastBoostersOpened.push([...action.payload])
-    },
-    resetLastBoostersOpened: (state) => {
-      state.lastBoostersOpened = []
     },
     resetLobby: () => initialState,
     addTournament: (state, action: PayloadAction<TournamentSchema>) => {
@@ -226,8 +214,6 @@ export const lobbySlice = createSlice({
 
 export const {
   removeMessage,
-  setBoosterContent,
-  resetLastBoostersOpened,
   pushMessage,
   setLeaderboard,
   setBotLeaderboard,
