+++ app/public/src/stores/index.ts
@@ -1,6 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit"
 import { enableMapSet } from "immer"
 import afterReducer from "./AfterGameStore"
+import boostersReducer from "./BoostersStore"
 import gameReducer from "./GameStore"
 import lobbyReducer from "./LobbyStore"
 import networkReducer from "./NetworkStore"
@@ -11,6 +12,7 @@ enableMapSet()
 const store = configureStore({
   reducer: {
     lobby: lobbyReducer,
+    boosters: boostersReducer,
     network: networkReducer,
     preparation: preparationReducer,
     game: gameReducer,
