+++ app/types/index.ts
@@ -65,8 +65,6 @@ export enum Transfer {
   NEW_MESSAGE = "NEW_MESSAGE",
   CHANGE_NAME = "CHANGE_NAME",
   CHANGE_AVATAR = "CHANGE_AVATAR",
-  OPEN_BOOSTER = "OPEN_BOOSTER",
-  BUY_BOOSTER = "BUY_BOOSTER",
   ADD_BOT = "ADD_BOT",
   REMOVE_BOT = "REMOVE_BOT",
   TOGGLE_READY = "TOGGLE_READY",
@@ -87,7 +85,6 @@ export enum Transfer {
   CHANGE_ROOM_RANKS = "CHANGE_ROOM_RANKS",
   CHANGE_SPECIAL_RULE = "CHANGE_SPECIAL_RULE",
   BUY_EMOTION = "BUY_EMOTION",
-  BOOSTER_CONTENT = "BOOSTER_CONTENT",
   USER = "USER",
   DRAG_DROP_CANCEL = "DRAG_DROP_CANCEL",
   SHOW_EMOTE = "SHOW_EMOTE",
